requirements.txt
streamlit
transformers
torch
sentencepiece
gspread
google-auth
